% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sl4.r
\name{read_SL4}
\alias{read_SL4}
\title{Read an SL4 solution file}
\usage{
read_SL4(filename, toLowerCase = TRUE)
}
\arguments{
\item{filename}{Path to SL4 file}
}
\value{
A list of variables
}
\description{
This function reads SL4 solution files into a list
}
