% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_har.R
\name{write_har}
\alias{write_har}
\title{Write an R list to GEMPACK HAR}
\usage{
write_har(data, filename)
}
\arguments{
\item{filename}{Path to HAR file}

\item{useCoefficientsAsNames}{If a coefficient name is present in the header, use that instead of the four-letter header}
}
\value{
A list of headers
}
\description{
Write an R list to GEMPACK HAR
}
