% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_har}
\alias{read_har}
\title{Read a GEMPACK HAR file into R}
\usage{
read_har(filename, useCoefficientsAsNames = F)
}
\arguments{
\item{filename}{Path to HAR file}
}
\value{
A list of headers
}
\description{
Reads in a GEMPACK HAR file and returns its  representation a list. Currently can only process integer headers, real full headers and character headers
}
