% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_har.R
\name{write_har}
\alias{write_har}
\title{Write an R list to GEMPACK HAR}
\usage{
write_har(data, filename, maxSize = 1e+07)
}
\arguments{
\item{filename}{Path to HAR file}
}
\value{
A list of headers
}
\description{
This function writes a HAR file based on a list
}
